feat: Add saveImmediate method to CollectionReference and DocumentReference
The code changes add a new method called `saveImmediate` to both the `CollectionReference` and `DocumentReference` classes. This method allows the Alwatr Store to save the collection or document immediately, bypassing any debounce delay. The purpose of this change is to provide a way to save the data without any delay when immediate saving is required.