feat: Open document and collection with given id in AlwatrStore
This commit adds the `openDocument` and `openCollection` methods to the `AlwatrStore` class. These methods allow users to open a document or collection with the given id and create and return a corresponding `DocumentReference` or `CollectionReference`. The previous `doc` and `collection` methods have been deprecated and will be removed in future versions.

BREAKING CHANGE: The `doc` and `collection` methods have been deprecated and should no longer be used. Instead, use the `openDocument` and `openCollection` methods.