refactor: rename `deleteFile` method to `remove`
This commit updates the `AlwatrStore` class by renaming the `deleteFile` method to `remove` to better reflect its purpose. The method removes a document or collection from the store and deletes the corresponding file from disk. If the file is not found, an error is thrown. If the file is not unloaded, it will be unloaded first. The previous method name was misleading and has been updated to provide clarity.

BREAKING CHANGE: The `deleteFile` method has been renamed to `remove`. Update your code accordingly.