feat: Add Alwatr Store package
This commit adds the Alwatr Store package, which is an extremely fast and compact JSON-based database that operates in memory. It includes a JSON file backup and serves over the highly accelerated Nginx. The package provides a convenient and efficient solution for data storage and retrieval. For more information, please visit [Alwatr Store](https://github.com/Alwatr/store#readme).