feat: Add getStoreList method to AlwatrStore
This refactor adds a new method, getStoreList, to the AlwatrStore class. The getStoreList method retrieves all store files from the database and returns them as an array of CollectionItem objects. This provides a convenient way to access and iterate over all the store files in the database. The method is implemented by returning the values of the rootDb__ object, which contains all the store files. This change improves the functionality and usability of the AlwatrStore class.