feat: Add migrateName property to StoreFileStat
This refactor adds a new property, migrateName, to the StoreFileStat type in the main.ts file. The migrateName property is used to specify the name of the migration process for migrating the store file to a new schema version. This change enhances the flexibility and extensibility of the StoreFileStat type, allowing for easier management of schema migrations in the database.