docs: update package descriptions for Nitrobase
Updated the package descriptions in the `package.json` files of `alwatr-nitrobase`, `@alwatr/nitrobase-types`, `@alwatr/nitrobase-engine`, `@alwatr/nitrobase-reference` to accurately reflect the features and benefits of Nitrobase. The new descriptions highlight Nitrobase as a blazingly fast, lightweight database built on JSON, with in-memory storage, JSON file backup, and the ability to serve data over the web using Nginx. This change improves the clarity and consistency of the package descriptions across the Nitrobase project.